fix: reject NaN/Inf usage, lock set_max_capacity, fix per_seconds type
- Validate usage values with math.isfinite() to prevent NaN/Inf from
  corrupting bucket state permanently
- Acquire backend lock in memory set_max_capacity to prevent torn
  reads of max_capacity/_rate_per_sec under concurrency
- Fix per_seconds type annotation float→int in callback Protocols and
  loguru implementations to match the Quota model migration

+++ token_throttle/_interfaces/_callbacks.py
@@ -67,7 +67,7 @@ class OnMissingConsumptionDataCallback(Protocol):
         *,
         model_family: str,
         usage_metric: str,
-        per_seconds: float,
+        per_seconds: int,
     ) -> None:
         """Called when no previous consumption data is detected, assuming full quota"""
 
@@ -197,7 +197,7 @@ def create_loguru_callbacks(
         *,
         model_family: str,
         usage_metric: str,
-        per_seconds: float,
+        per_seconds: int,
     ) -> None:
         logger = _get_loguru_logger()
         logger.log(
@@ -284,7 +284,7 @@ class SyncOnMissingConsumptionDataCallback(Protocol):
         *,
         model_family: str,
         usage_metric: str,
-        per_seconds: float,
+        per_seconds: int,
     ) -> None: ...
 
 
@@ -401,7 +401,7 @@ def create_sync_loguru_callbacks(
         *,
         model_family: str,
         usage_metric: str,
-        per_seconds: float,
+        per_seconds: int,
     ) -> None:
         logger = _get_loguru_logger()
         logger.log(
